+++ jest.config.ts
@@ -1,12 +1,11 @@
 import type { Config } from 'jest';
+import { nestEsmTransform } from './test/jest-nest-esm';
 
 const config: Config = {
+  ...nestEsmTransform,
   moduleFileExtensions: ['js', 'json', 'ts'],
   rootDir: '.',
   testRegex: 'test/(?!e2e/).*\\.spec\\.ts$',
-  transform: {
-    '^.+\\.ts$': 'ts-jest',
-  },
   collectCoverageFrom: ['src/**/*.ts', '!src/index.ts'],
   coverageDirectory: './coverage',
   coverageThreshold: {