+++ test/e2e/jest-e2e.config.ts
@@ -1,12 +1,11 @@
 import type { Config } from 'jest';
+import { nestEsmTransform } from '../jest-nest-esm';
 
 const config: Config = {
+  ...nestEsmTransform,
   moduleFileExtensions: ['js', 'json', 'ts'],
   rootDir: '../..',
   testRegex: 'test/e2e/.*\\.e2e-spec\\.ts$',
-  transform: {
-    '^.+\\.ts$': 'ts-jest',
-  },
   testEnvironment: 'node',
 };